requirements.txt
# Core dependencies
streamlit==1.28.0
numpy==1.24.3
Pillow==10.0.1

# Computer Vision - specific versions that work together
opencv-python-headless==4.8.1.78

# PyTorch - CPU only to avoid CUDA issues
torch==2.1.0
torchvision==0.16.0

# YOLO
ultralytics==8.0.196

# Additional dependencies that might be missing
scipy==1.11.3
matplotlib==3.7.2
requests==2.31.0